weeb_api/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==6.0.2
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
